# packages/store/src/types.ts
/** JSON-safe value. Decoded payloads are always representable in plain JSON. */
export type JsonValue = string | number | boolean | null | JsonValue[] | { [k: string]: JsonValue };

/**
 * Module 1's `RawEvent`, redeclared structurally.
 *
 * Module 2 does not import `@soroban-lens/ingest`: keeping the two packages
 * free of a build dependency is what lets Person A and Person B work in
 * parallel. The shape must stay identical — see CONTRIBUTING.md.
 */
export interface RawEventInput {
  id: string;
  type: 'contract' | 'system';
  ledger: number;
  ledgerClosedAt: string;
  contractId: string;
  topic: string[];
  value: string;
  txHash: string;
  transactionIndex: number;
  operationIndex: number;
  inSuccessfulContractCall: boolean;
}

/** One decoded XDR `ScVal`. */
export interface DecodedValue {
  /** ScVal arm, lower-cased: "symbol", "i128", "address", "map", "vec", "bytes", ... */
  type: string;
  /** JSON-safe rendering. BigInts become decimal strings, bytes become hex. */
  value: JsonValue;
}

/**
 * A stored, decoded event. This is the record the API serves and the UI renders.
 *
 * Raw XDR is kept alongside the decoding so that a decoder bug is always
 * recoverable: re-run the decode over `topicsXdr` / `valueXdr` without
 * re-indexing from the network.
 */
export interface LensEvent {
  id: string;
  contractId: string;
  type: 'contract' | 'system';
  ledger: number;
  /** RFC3339, exactly as the RPC reported it. */
  ledgerClosedAt: string;
  txHash: string;
  transactionIndex: number;
  operationIndex: number;
  inSuccessfulContractCall: boolean;
  /** Full decoded topic list. May be longer than 4 — see fixtures/README.md. */
  topics: DecodedValue[];
  /** Original base64 XDR topics. */
  topicsXdr: string[];
  value: DecodedValue;
  /** Original base64 XDR value. */
  valueXdr: string;
  /** Set when XDR decoding failed; the raw fields are still populated. */
  decodeError?: string | undefined;
  /** When this row was written, ISO 8601. */
  indexedAt: string;
}

/** Filter for `EventStore.queryEvents`. All fields are ANDed. */
export interface EventQuery {
  contractId?: string | undefined;
  /**
   * Match a topic prefix: `['transfer']` matches any event whose first topic is
   * `transfer`. A `null` entry is a wildcard for that position.
   * Prefix rather than exact match, because events may carry more topics than
   * a filter can name.
   */
  topics?: (string | null)[] | undefined;
  /** Inclusive lower bound on ledger sequence. */
  fromLedger?: number | undefined;
  /** Inclusive upper bound on ledger sequence. */
  toLedger?: number | undefined;
  txHash?: string | undefined;
  /** Restrict to successful contract calls. Omitted means "both". */
  successfulOnly?: boolean | undefined;
  /** 1..1000. Defaults to 50. */
  limit?: number | undefined;
  /** Keyset cursor: return events strictly before/after this event id. */
  cursor?: string | undefined;
  /** "desc" (newest first, the default) or "asc". */
  order?: 'asc' | 'desc' | undefined;
}

/** A page of query results plus the token for the next page. */
export interface EventPage {
  events: LensEvent[];
  /** Pass back as `cursor` for the next page. `null` when the page is the last. */
  nextCursor: string | null;
  /** Total rows matching the filter, ignoring limit/cursor. */
  total: number;
}

/** Aggregate view of one indexed contract. */
export interface ContractSummary {
  contractId: string;
  eventCount: number;
  firstLedger: number;
  lastLedger: number;
  lastSeenAt: string;
}

/** Progress of one ingest stream, mirrored into storage so the API can report it. */
export interface StreamState {
  key: string;
  cursor: string;
  ledger: number;
  updatedAt: string;
}

export interface StoreStats {
  eventCount: number;
  contractCount: number;
  minLedger: number | null;
  maxLedger: number | null;
  /** Schema version the database is currently migrated to. */
  schemaVersion: number;
}